renamed user.ts -> userSession.ts, added user.ts to post user to database
